requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
cfgv==3.4.0
distlib==0.3.7
Django==4.2.5
django-js-asset==2.1.0
django-mptt==0.14.0
django-mptt-admin==2.5.1
filelock==3.12.4
identify==2.5.30
nodeenv==1.8.0
platformdirs==3.11.0
pre-commit==3.4.0
python-dotenv==1.0.0
PyYAML==6.0.1
soupsieve==2.5
sqlparse==0.4.4
typing-extensions==4.8.0
virtualenv==20.24.5
